+++ src/caching/Arweave/index.ts
@@ -1,2 +1,1 @@
 export * from "./ArweaveClient";
-export * from "./types";
