+++ src/interfaces/index.ts
@@ -7,3 +7,4 @@ export * from "./UBA";
 export * from "./Error";
 export * from "./TypedData";
 export * from "./CachingMechanism";
+export * from "./Caching";
